fix/comments
