+++ src/components/quota-configuration/edit-quota-configuration-modal.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
@@ -15,9 +15,9 @@ import {
 } from '@/components/ui/dialog';
 import { QuotaConfiguration } from '@/types/quota-configuration.type';
 import { Edit } from 'lucide-react';
-import { useClientDate } from '@/hooks/use-client-date';
+import { useEditQuotaConfiguration } from '../../hooks/quota-configuration/use-edit-quota-configuration';
 
-interface EditQuotaConfigurationModalProps {
+interface Props {
     configuration: QuotaConfiguration;
     onSubmit: (id: string, data: { fee?: number; start_date?: string }) => Promise<void>;
     loading?: boolean;
@@ -27,112 +27,44 @@ export function EditQuotaConfigurationModal({
     configuration,
     onSubmit,
     loading = false
-}: EditQuotaConfigurationModalProps) {
+}: Props) {
     const [open, setOpen] = useState(false);
-    const [formData, setFormData] = useState({
-        fee: configuration.fee.toString(),
-        start_date: '',
-    });
-    const [errors, setErrors] = useState<Record<string, string>>({});
-    const [canEdit, setCanEdit] = useState(true);
-    const [editReason, setEditReason] = useState<string>('');
-    const currentDate = useClientDate();
 
-    useEffect(() => {
-        if (!currentDate) return;
-
-        // Validar si se puede editar según las reglas del backend
-        const startDate = new Date(configuration.start_date);
-        const lastDayAbleToUpdate = new Date(startDate);
-        lastDayAbleToUpdate.setDate(lastDayAbleToUpdate.getDate() + 7);
-
-        const tomorrow = new Date(currentDate);
-        tomorrow.setDate(tomorrow.getDate() + 1);
-
-        // Verificar si pasaron más de 7 días
-        if (lastDayAbleToUpdate < currentDate) {
-            setCanEdit(false);
-            setEditReason('No se puede editar después de 7 días desde la fecha de inicio');
-            return;
-        }
-
-        // Verificar si es la primera configuración sin fecha de fin
-        if (configuration.quota_amount_configuration_number === 1 && !configuration.end_date) {
-            setCanEdit(false);
-            setEditReason('No se puede editar la primera configuración activa');
-            return;
-        }
-
-        setCanEdit(true);
-        setEditReason('');
-    }, [configuration, currentDate]);
-
-    const validateForm = () => {
-        const newErrors: Record<string, string> = {};
-
-        if (formData.fee && (isNaN(Number(formData.fee)) || Number(formData.fee) <= 0)) {
-            newErrors.fee = 'La cuota debe ser un número positivo';
-        }
-
-        if (formData.start_date && currentDate) {
-            const selectedDate = new Date(formData.start_date + 'T00:00:00');
-            const tomorrow = new Date(currentDate);
-            tomorrow.setDate(tomorrow.getDate() + 1);
-            tomorrow.setHours(0, 0, 0, 0);
-
-            if (selectedDate > tomorrow) {
-                newErrors.start_date = 'La fecha de inicio no puede ser posterior a mañana';
-            }
-        }
-
-        setErrors(newErrors);
-        return Object.keys(newErrors).length === 0;
-    };
+    const {
+        formData,
+        errors,
+        canEdit,
+        editReason,
+        handleInputChange,
+        validateForm,
+        resetForm,
+    } = useEditQuotaConfiguration(configuration);
 
     const handleSubmit = async (e: React.FormEvent) => {
         e.preventDefault();
+        if (!validateForm()) return;
 
-        if (!validateForm()) {
-            return;
+        const updateData: { fee?: number; start_date?: string } = {};
+
+        if (formData.fee && Number(formData.fee) !== configuration.fee) {
+            updateData.fee = Number(formData.fee);
         }
 
-        try {
-            const updateData: { fee?: number; start_date?: string } = {};
-
-            if (formData.fee && Number(formData.fee) !== configuration.fee) {
-                updateData.fee = Number(formData.fee);
-            }
-
-            if (formData.start_date) {
-                const dateWithTime = new Date(formData.start_date + 'T00:00:00');
-                updateData.start_date = dateWithTime.toISOString();
-            }
-
-            // Solo enviar si hay cambios
-            if (Object.keys(updateData).length > 0) {
-                await onSubmit(configuration._id, updateData);
-                setOpen(false);
-            }
-        } catch {
-            // El error se maneja en el componente padre
+        if (formData.start_date) {
+            const dateWithTime = new Date(formData.start_date + 'T00:00:00');
+            updateData.start_date = dateWithTime.toISOString();
         }
-    };
 
-    const handleInputChange = (field: string, value: string) => {
-        setFormData(prev => ({ ...prev, [field]: value }));
-        if (errors[field]) {
-            setErrors(prev => ({ ...prev, [field]: '' }));
+        if (Object.keys(updateData).length > 0) {
+            await onSubmit(configuration._id, updateData);
+            setOpen(false);
         }
     };
 
     const handleOpenChange = (newOpen: boolean) => {
         setOpen(newOpen);
         if (!newOpen) {
-            setFormData({
-                fee: configuration.fee.toString(),
-                start_date: '',
-            });
-            setErrors({});
+            resetForm();
         }
     };
 
@@ -215,4 +147,4 @@ export function EditQuotaConfigurationModal({
             </DialogContent>
         </Dialog>
     );
-} 
+}
