+++ src/types/sidebar.types.ts
@@ -1,0 +1,8 @@
+import { LucideIcon } from "lucide-react";
+
+export interface MenuItems {
+    title: string;
+    url: string;
+    icon: LucideIcon;
+    isUserMenu?: boolean
+}
