+++ src/types/user.type.ts
@@ -1,0 +1,8 @@
+export interface CurrentUser {
+    _id: string;
+    username: string;
+    role: {
+        _id: string;
+        name: string;
+    };
+};
